% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelled_spss_survey.R
\name{as_factor}
\alias{as_factor}
\title{Convert labelled_spss_survey vector To Factor}
\usage{
as_factor(x, levels = "defaults")
}
\arguments{
\item{x}{A vector to label. Must be either numeric (integer or double) or
character.}

\item{levels}{How to create the levels of the generated factor:
\itemize{
\item "default": uses labels where available, otherwise the values.
Labels are sorted by value.
\item "both": like "default", but pastes together the level and value
\item "label": use only the labels; unlabelled values become \code{NA}
\item "values: use only the values
}}
}
\description{
Convert a \code{\link{labelled_spss_survey}} vector to a type 
of factor.
}
\seealso{
\code{as_factor} is imported from \code{haven::\link[haven:as_factor]{as_factor}}
}
