% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codelist.R
\name{codelist_create}
\alias{codelist_create}
\title{Create a codelist}
\usage{
codelist_create()
}
\value{
This function will return a data frame (tbl_df) with a codelist.
}
\description{
This is a function candidate.
}
\details{
should be a new function that creates a codelist which considers  
\href{SDMX Content-Oriented Guidelines (COG)}{https://sdmx.org/?page_id=4345}
and any guidance from DDI on Question banks, and the DDI Question Construct.
Partly replaces the current \code{\link{create_codebook}} function.
}
\examples{
# This is a new function candidate and is not written yet.
codelist_create()
}
\concept{codelist functions}
