% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_to_labelled_spss}
\alias{convert_to_labelled_spss}
\title{Convert to haven_labelled_spss}
\usage{
convert_to_labelled_spss(x, na_labels = NULL)
}
\arguments{
\item{x}{A vector}

\item{na_labels}{A named vector of missing values, defaults to
\code{c( "inap" = "inap")} for character vectors and
\code{c( 99999 = "inap")} for numeric vectors.}
}
\value{
A haven_labelled_spss vector
}
\description{
Convert to haven_labelled_spss
}
\keyword{internal}
