% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_waves.R
\name{document_waves}
\alias{document_waves}
\title{Document survey lists}
\usage{
document_waves(survey_list)
}
\arguments{
\item{survey_list}{A list of \code{\link{survey}} objects.}
}
\value{
Returns a data frame with the key attributes of the surveys in
a survey list.
}
\description{
Document survey lists
}
\examples{
survey_list <- dir (
    here( "inst", "examples"))[grepl(".rds", 
                        dir (here( "inst", "examples")))]
                        
example_surveys <- read_surveys(
    here( "inst", "examples", survey_list))
    
waves_document <- document_waves(example_surveys)

attr(waves_document, "original_list" )
waves_document
}
\seealso{
Other documentation functions: 
\code{\link{document_survey_item}()}
}
\concept{documentation functions}
