% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_waves.R
\name{harmonize_waves}
\alias{harmonize_waves}
\title{Harmonize waves}
\usage{
harmonize_waves(waves, .f)
}
\arguments{
\item{waves}{A list of surveys}

\item{.f}{A function}
}
\description{
Harmonize the values of surveys. Not fully functional yet.
}
