% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_range_to_values.R
\name{is.na_range_to_values}
\alias{is.na_range_to_values}
\title{Test whether missing values need harmonization}
\usage{
is.na_range_to_values(x)
}
\arguments{
\item{x}{A labelled vector.}
}
\value{
Logical scalar.
}
\description{
Checks whether both `na_values` and `na_range` attributes
are present on a labelled vector.
}
\keyword{internal}
