% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_surveys.R
\name{merge_waves}
\alias{merge_waves}
\title{Deprecated wrapper for `merge_surveys()`}
\usage{
merge_waves(waves, var_harmonization)
}
\arguments{
\item{waves}{Deprecated alias for `survey_list`.}

\item{var_harmonization}{A metadata table describing the
  harmonization rules. The table must contain at least:

  - `filename`
  - `var_name_orig`
  - `var_name_target`
  - `var_label`}
}
\value{
A list of harmonized survey objects.
}
\description{
`merge_waves()` has been renamed to [merge_surveys()]
for more general survey harmonization workflows.
}
\seealso{
[merge_surveys()]

Other survey harmonization functions:
\code{\link[=merge_surveys]{merge_surveys()}}
}
\concept{survey harmonization functions}
