% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_create.R
\name{metadata_waves_create}
\alias{metadata_waves_create}
\title{Create a metadata table}
\usage{
metadata_waves_create(survey_list)
}
\arguments{
\item{survey_list}{A list containing surveys of class survey.}
}
\description{
Create a metadata table
}
\examples{
examples_dir <- system.file( "examples", package = "retroharmonize")

my_rds_files <- dir( examples_dir)[grepl(".rds", 
                                        dir(examples_dir))]

example_surveys <- read_surveys(file.path(examples_dir, my_rds_files))
metadata_waves_create (example_surveys)
}
\seealso{
Other metadata functions: 
\code{\link{codebook_create}()},
\code{\link{metadata_create}()}
}
\concept{metadata functions}
