% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_create.R
\name{metadata_initialize}
\alias{metadata_initialize}
\title{Initialize a metadata data frame}
\usage{
metadata_initialize(filename, id)
}
\value{
A nested data frame with metadata and the range of 
labels, na_values and the na_range itself.
}
\description{
Initialize a metadata data frame
}
\keyword{internal}
