% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retroharmonize.R
\docType{package}
\name{retrohamonize}
\alias{retrohamonize}
\title{retroharmonize: A package to create tidy Eurobarometer microdata files.}
\description{
The goal of retroharmonize is to allow the organization of data 
joins or panels from various data sources, particularly survey 
microdata files, by retrospective harmonization the value codes, 
the value labels, and the missing value ranges of the data in a 
reproducible manner with the help of comprehensive s3 classes.
}
\section{import functions}{

The naming functions make the GESIS SPSS files usable in a 
programmatic context.
}

