% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_surveys.R
\name{subset_survey_file}
\alias{subset_survey_file}
\title{Subset surveys from files}
\usage{
subset_survey_file(
  file_path,
  subset_vars,
  subset_name = "subset",
  id = NULL,
  export_path = NULL
)
}
\arguments{
\item{file_path}{A single survey files.}

\item{subset_vars}{Character vector of variable names to retain.
If `NULL`, all variables are retained.}

\item{subset_name}{Character string appended to filenames of
subsetted surveys. Defaults to `"subset"`.}

\item{export_path}{Optional directory where subsetted surveys
are saved as `.rds` files. If `NULL`, surveys are returned in memory.}
}
\description{
Subset surveys from files
}
\keyword{internal}
