% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_surveys.R
\name{subset_survey_files}
\alias{subset_survey_files}
\title{Subset surveys from files}
\usage{
subset_survey_files(
  file_path,
  subset_vars,
  subset_name = "subset",
  id = NULL,
  export_path = NULL
)
}
\arguments{
\item{file_path}{A single survey files.}

\item{subset_vars}{The names of the variables that should be kept from all surveys in the list that contains the
wave of surveys. Defaults to \code{NULL} in which case it returns all variables without subsetting.}

\item{subset_name}{An identifier for the survey subset.}
}
\description{
Subset surveys from files
}
\keyword{internal}
