% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggest_var_names.R
\name{suggest_permanent_names}
\alias{suggest_permanent_names}
\title{Suggest Permanent Names}
\usage{
suggest_permanent_names(survey_program = "eurobarometer")
}
\arguments{
\item{survey_program}{Suggest permanent names for the survey progarm \code{"eurobarometer"}}
}
\value{
A character vector with suggested permanent names.
}
\description{
Established survey programs usually have their own variable name conventions. 
The suggested constant names keep these variable names constant.
}
\examples{
suggest_permanent_names ( "eurobarometer" )
}
