% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveyharmonize.R
\docType{package}
\name{surveyhamonize}
\alias{surveyhamonize}
\title{surveyharmonize: A package to create tidy Eurobarometer microdata files.}
\description{
The package eurobarometer has a number of related functions to
analyse metadata; labelling variables with the help of metadata;
harmonizing variable names, classes and labels.
}
\section{joining functions}{

The naming functions make the GESIS SPSS files usable in a programmatic
context.
}

