% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{validate_harmonize_labels}
\alias{validate_harmonize_labels}
\title{Validate harmonize_labels parameter
Check if "from", "to", and "numeric_values" are of equal lengths.}
\usage{
validate_harmonize_labels(harmonize_labels)
}
\description{
Validate harmonize_labels parameter
Check if "from", "to", and "numeric_values" are of equal lengths.
}
\keyword{internal}
