% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelled_spss_survey.R
\name{as_factor}
\alias{as_factor}
\title{Convert labelled_spss_survey vector To Factor}
\usage{
as_factor(x, ...)
}
\arguments{
\item{x}{Object to coerce to a factor.}

\item{...}{Other arguments passed down to method.}
}
\description{
Convert a \code{\link{labelled_spss_survey}} vector to a type 
of factor.
}
\seealso{
\code{as_factor} is imported from \code{haven::\link[haven:as_factor]{as_factor}}
}
