% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_labelled_spss_survey.R
\name{as_labelled_spss_survey}
\alias{as_labelled_spss_survey}
\title{Labelled to labelled_spss_survey}
\usage{
as_labelled_spss_survey(x, id)
}
\arguments{
\item{x}{A vector of class haven_labelled or haven_labelled_spss.}

\item{id}{The survey identifier.}
}
\value{
A vector of labelled_spss_survey
}
\description{
Labelled to labelled_spss_survey
}
\seealso{
Other type conversion functions: 
\code{\link{labelled_spss_survey}()}
}
\concept{type conversion functions}
