% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\name{codebook_create}
\alias{codebook_create}
\title{Create a codebook}
\usage{
codebook_create()
}
\value{
This function will return
}
\description{
This is a function candidate which will partly replace the current
\code{\link{create_codebook}} function.
}
\details{
This future function should create a DDI-Codebook compatible, partial codebook on 
survey level.
}
\examples{
# This is a new function candidate and is not written yet.
codebook_create()
}
\concept{codebook functions}
\keyword{internal}
