% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook_create.R
\name{codebook_create}
\alias{codebook_create}
\title{Create a codebook}
\usage{
codebook_create(metadata, survey = NULL)
}
\arguments{
\item{metadata}{A metadata table created by \code{\link{metadata_create}}.}

\item{survey}{A survey data frame, defaults to \code{NULL}. If the survey is
given as parameter, the metadata will be set to the metadata of this particular
survey by \code{\link{metadata_create}}.}
}
\value{
A codebook for the survey as a data frame, including the metadata, and all found 
SPSS-type valid or missing labels.
}
\description{
Create a metadata table from the survey data files.
}
\examples{
codebook_create (
 survey = read_rds (
          system.file("examples", "ZA7576.rds",
                      package = "retroharmonize")
          )
)
}
