% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosswalk.R
\name{crosswalk_surveys}
\alias{crosswalk_surveys}
\alias{crosswalk}
\title{Crosswalk surveys}
\usage{
crosswalk_surveys(survey_list, crosswalk_table, na_values = NULL)

crosswalk(survey_list, crosswalk_table, na_values = NULL)
}
\arguments{
\item{survey_list}{A list of \code{\link{survey}} objects.}

\item{crosswalk_table}{A table created with \code{crosswalk_table_create}, or a 
data frame with at least the following columns: \code{var_name_orig}, \code{var_name_target},
for harmonizing the variable names. If \code{val_label_orig}, \code{val_label_target} 
are present, the value labels will be harmonized, too. If 
\code{var_numeric_orig}, \code{var_numeric_target} are present, the numeric codes of the 
variable will be harmonized. If \code{class_target} is present, then the class of the variable 
will be harmonized to any of \code{factor}, \code{numeric} or \code{character} using 
\code{\link{as_factor}}, \code{\link{as_numeric}}, or \code{\link{as_character}}.}

\item{na_values}{A named vector of \code{na_values}, the 
observations that are defined to be treated as missing in 
the SPSS-style coding. Defaults to \code{NULL}.}
}
\value{
\code{crosswalk} will return a data frame, and \code{crosswalk_surveys} a list of
data frames, where the variable names, and optionally the variable labels, and the missing 
value range is harmonized (the same names, labels, codes are used.)
}
\description{
Harmonize surveys with crosswalk tables.
}
\details{
Harmonize a survey or a list of surveys with the help of a crosswalk table.
You can create the crosswalk table with \code{\link{crosswalk_table_create}}, or manually 
create a crosswalk table as a data frame including at least the following columns: \code{id}
for identifying a survey, \code{var_name_orig} for the original variable name 
and \code{var_name_target} for the new (target) variable name. Optionally you can harmonize
the value labels, the numeric codes, and the special missing labels, too.
}
\examples{
\donttest{
examples_dir <- system.file("examples", package = "retroharmonize")
survey_list <- dir(examples_dir)[grepl("\\\\.rds", dir(examples_dir))]
example_surveys <- read_surveys(
  file.path( examples_dir, survey_list), 
  save_to_rds = FALSE)

## Compare with documentation:
documented_surveys <- metadata_surveys_create(example_surveys)
documented_surveys <- documented_surveys[
documented_surveys$var_name_orig \%in\% c( "rowid", "isocntry", "w1", "qd3_4",
                                          "qd3_8" , "qd7.4", "qd7.8", "qd6.4", "qd6.8"),
                                          ]
crosswalk_table <- crosswalk_table_create ( metadata = documented_surveys )
}
}
\seealso{
Other harmonization functions: 
\code{\link{collect_val_labels}()},
\code{\link{crosswalk_table_create}()},
\code{\link{harmonize_na_values}()},
\code{\link{harmonize_survey_values}()},
\code{\link{harmonize_values}()},
\code{\link{harmonize_var_names}()},
\code{\link{label_normalize}()},
\code{\link{suggest_permanent_names}()},
\code{\link{suggest_var_names}()}
}
\concept{harmonization functions}
