% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosswalk.R
\name{crosswalk_table_create}
\alias{crosswalk_table_create}
\title{Create a crosswalk table}
\usage{
crosswalk_table_create(metadata)
}
\arguments{
\item{metadata}{A metadata table created by \code{\link{metadata_create}}.}
}
\value{
A tibble with raw crosswalk table. It contains all harmonization tasks, but the 
target values need to be set by further manipulations.
}
\description{
Create a crosswalk table with the source variable names and variable labels.
}
\details{
The table contains a \code{var_name_target} and \code{val_label_target} column, but 
these values need to be set by further manual or reproducible harmonization steps.
}
