% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_na_values.R
\name{harmonize_na_values}
\alias{harmonize_na_values}
\title{Harmonize na_values in haven_labelled_spss}
\usage{
harmonize_na_values(df)
}
\arguments{
\item{df}{A data frame that contains haven_labelled_spss vectors.}
}
\value{
A tibble where the na_values are consistent
}
\description{
Harmonize na_values in haven_labelled_spss
}
\examples{
\dontrun{
data ( "ZA6863_sample", package="eurobarometer")

harmonize_na_values(ZA6863_sample)
}
}
\seealso{
Other harmonization functions: 
\code{\link{harmonize_values}()},
\code{\link{harmonize_waves}()},
\code{\link{merge_waves}()}
}
\concept{harmonization functions}
