% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_survey_variables.R
\name{harmonize_survey_variables}
\alias{harmonize_survey_variables}
\title{Harmonize survey variables}
\usage{
harmonize_survey_variables(
  crosswalk_table,
  subset_name = "subset",
  survey_list = NULL,
  survey_paths = NULL,
  import_path = NULL,
  export_path = NULL
)
}
\arguments{
\item{crosswalk_table}{A crosswalk table created by \code{\link{crosswalk_table_create}} or a manually created 
crosstable including at least
\code{filename}, \code{var_name_orig}, \code{var_name_target} and optionally 
 \code{var_label_orig} and \code{var_label_target}. This parameter is optional and 
 defaults to \code{NULL}.}

\item{subset_name}{An identifier for the survey subset.}

\item{survey_list}{A list of surveys imported with \code{\link{read_surveys}}. If set to 
\code{NULL}, the \code{survey_path} should give full path to the surveys.}

\item{survey_paths}{A vector of full file paths to the surveys to subset.}
}
\value{
A list of surveys or save individual rds files on the \code{export_path}.
}
\description{
Similar to \code{\link{subset_surveys}}, but will not only remove the 
variables that cannot be harmonized, but also renames the variables that are kept.
}
\examples{
{
examples_dir <- system.file("examples", package = "retroharmonize")
survey_list  <- dir(examples_dir)[grepl("\\\\.rds", dir(examples_dir))]
example_surveys <- read_surveys(
  file.path( examples_dir, survey_list), 
  export_path = NULL)

documented_surveys <- metadata_create(example_surveys)
documented_surveys <- documented_surveys[
   documented_surveys$var_name_orig \%in\% c( "rowid", "isocntry", "w1", "qd3_4",
   "qd3_8" , "qd7.4", "qd7.8", "qd6.4", "qd6.8"),
   ]
crosswalk_table    <- crosswalk_table_create ( metadata = documented_surveys )

freedom_table <- crosswalk_table[
                     which(crosswalk_table$var_name_target \%in\% c("rowid", "freedom")),
                     ]

harmonize_survey_variables(crosswalk_table = freedom_table,
                           subset_name     = 'freedom',
                           survey_list     = example_surveys )

}
}
