% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_values.R
\name{harmonize_values}
\alias{harmonize_values}
\title{Harmonize the values and labels of labelled vectors}
\usage{
harmonize_values(
  x,
  harmonize_labels = list(from = c("^tend\\\\sto|^trust",
    "^tend\\\\snot|not\\\\strust", "^dk", "^inap"), to = c("trust", "not_trust",
    "do_not_know", "inap"), numeric_value = c(1, 0, 99997, 99999)),
  na_values = c("do_not_know", "inap"),
  id = "survey_id"
)
}
\arguments{
\item{x}{A labelled vector}

\item{harmonize_labels}{A list of harmonization values}

\item{na_values}{A named vector of na_values.}

\item{id}{A survey ID, defaults to \code{survey_id}}
}
\description{
Harmonize the values and labels of labelled vectors
}
\examples{
var1 <- labelled::labelled_spss(x = c(1,0,1,1,0,8,9), 
                                labels = c("TRUST" = 1, 
                                           "NOT TRUST" = 0, 
                                           "DON'T KNOW" = 8, 
                                           "INAP. HERE" = 9), 
                                na_values = c(8,9))

harmonize_values (var1)                                

var2 <- labelled::labelled_spss(x = c(1,6,2,9,1,1,2), 
                                labels = c("Tend to trust" = 1, 
                                           "Tend not to trust" = 2, 
                                           "DK" = 8, 
                                           "Inap" = 9), 
                                na_values = c(6,9))

harmonize_values (var2)

var3 <- labelled::labelled(x = c(1,6,2,9,1,1,2), 
                           labels = c("Tend to trust" = 1, 
                                      "Tend not to trust" = 2, 
                                      "DK" = 6))
harmonize_values (var3)
}
\seealso{
Other harmonization functions: 
\code{\link{harmonize_na_values}()}
}
\concept{harmonization functions}
