% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_waves.R
\name{harmonize_waves}
\alias{harmonize_waves}
\title{Harmonize waves}
\usage{
harmonize_waves(waves, .f)
}
\arguments{
\item{waves}{A list of surveys}

\item{.f}{A function to apply for the harmonization.}
}
\description{
Harmonize the values of surveys. It binds together variables
that are all present in the surveys, and applies a 
harmonization function on them.
}
\seealso{
Other harmonization functions: 
\code{\link{collect_val_labels}()},
\code{\link{harmonize_na_values}()},
\code{\link{harmonize_values}()},
\code{\link{merge_waves}()}
}
\concept{harmonization functions}
