% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_waves.R
\name{merge_waves}
\alias{merge_waves}
\title{Merge waves}
\usage{
merge_waves(waves, var_harmonization)
}
\arguments{
\item{waves}{A list of surveys}

\item{var_harmonization}{Metadata of surveys}
}
\value{
A list of surveys.
}
\description{
Merge a list of surveys into a list with harmonized variable names,
variable labels and survey identifiers.
}
\seealso{
survey
}
