% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_waves.R
\name{merge_waves}
\alias{merge_waves}
\title{Merge waves}
\usage{
merge_waves(waves, var_harmonization)
}
\arguments{
\item{waves}{A list of surveys}

\item{var_harmonization}{Metadata of surveys, including at least
\code{filename}, \code{var_name_orig}, \code{var_name}, \code{var_label}.}
}
\value{
A list of surveys with harmonized names and labels.
}
\description{
Merge a list of surveys into a list with harmonized variable names,
variable labels and survey identifiers.
}
\examples{
\dontrun{
require(dplyr)
survey_list <- dir (
    here( "inst", "examples"))[grepl(".rds", 
                        dir (here( "inst", "examples")))]
                        
example_surveys <- read_surveys(
    here( "inst", "examples", survey_list))
    
metadata <- lapply ( X = example_surveys, FUN = metadata_create )
metadata <- do.call(rbind, metadata)
 
to_harmonize <- metadata \%>\%
  filter ( var_name_orig \%in\% 
             c("rowid", "w1") |
             grepl("trust ", label_orig ) ) \%>\%
  mutate ( var_label = var_label_normalize(label_orig) ) \%>\%
  mutate ( var_name = val_label_normalize(var_label) )

merge_waves ( example_surveys, to_harmonize )
}
}
\seealso{
survey

Other harmonization functions: 
\code{\link{collect_val_labels}()},
\code{\link{harmonize_na_values}()},
\code{\link{harmonize_values}()},
\code{\link{harmonize_waves}()}
}
\concept{harmonization functions}
