% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_create.R
\name{metadata_create}
\alias{metadata_create}
\alias{metadata_waves_create}
\title{Create a metadata table from several surveys}
\usage{
metadata_create(survey_list = NULL, survey_paths = NULL, .f = NULL)

metadata_waves_create(survey_list)
}
\arguments{
\item{survey_list}{A list containing surveys of class survey.}

\item{survey_paths}{Optional character vector of file paths to surveys.}

\item{.f}{A function to import the surveys with.}
}
\description{
Create a metadata table from several surveys
}
\details{
The form \code{metadata_waves_create} is deprecated.
}
\examples{
examples_dir <- system.file("examples", package = "retroharmonize")

my_rds_files <- dir(examples_dir)[grepl(
  ".rds",
  dir(examples_dir)
)]

example_surveys <- read_surveys(file.path(examples_dir, my_rds_files))
metadata_create(example_surveys)
}
\seealso{
Other metadata functions: 
\code{\link{create_codebook}()},
\code{\link{is.crosswalk_table}()},
\code{\link{metadata_survey_create}()}
}
\concept{metadata functions}
