% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_create.R
\name{metadata_create}
\alias{metadata_create}
\title{Create a metadata table}
\usage{
metadata_create(survey)
}
\arguments{
\item{survey}{A survey data frame.}
}
\description{
Create a metadata table
}
\examples{
metadata_create (
 survey = read_rds (
          system.file("examples", "ZA7576.rds",
                      package = "retroharmonize")
          )
)
}
