% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_survey.R
\name{pull_survey}
\alias{pull_survey}
\title{Pull a survey from a survey list}
\usage{
pull_survey(survey_list, id = NULL, filename = NULL)
}
\arguments{
\item{survey_list}{A list of surveys}

\item{id}{The id of the requested survey. If \code{NULL} use 
\code{filename}}

\item{filename}{The filename of the requested survey.}
}
\value{
A single survey identified by \code{id} or \code{filename}.
}
\description{
Pull a survey by survey code or id.
}
\examples{
survey_list <- dir (
  here( "inst", "examples"))[grepl(".rds", 
                        dir (here( "inst", "examples")))]

example_surveys <- read_surveys(
  here( "inst", "examples", survey_list))
  
pull_survey(example_surveys, id = "ZA5913")
}
