% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_csv.R
\name{read_csv}
\alias{read_csv}
\title{Read csv file}
\usage{
read_csv(
  file,
  id = NULL,
  doi = NULL,
  header = FALSE,
  sep = "",
  quote = "\\"'",
  dec = ".",
  numerals = c("allow.loss", "warn.loss", "no.loss"),
  na.strings = "NA",
  skip = 0,
  check.names = TRUE,
  strip.white = FALSE,
  blank.lines.skip = TRUE,
  stringsAsFactors = FALSE,
  fileEncoding = "",
  encoding = "unknown"
)
}
\arguments{
\item{file}{A path to a file to import.}

\item{id}{An identifier of the tibble, if omitted, defaults to the
file name without its extension.}

\item{doi}{An optional document object identifier.}
}
\value{
A tibble, data frame variant with survey attributes.
}
\description{
Import a survey from a csv file.
}
\examples{
path <-  system.file("examples", "ZA7576.rds", package = "retroharmonize")
read_survey <- read_rds(path)
attr(read_survey, "id")
attr(read_survey, "filename")
attr(read_survey, "doi") 
}
\seealso{
Other import functions: 
\code{\link{pull_survey}()},
\code{\link{read_dta}()},
\code{\link{read_rds}()},
\code{\link{read_spss}()},
\code{\link{read_surveys}()}
}
\concept{import functions}
