% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_surveys.R
\name{read_surveys}
\alias{read_surveys}
\title{Read Survey Files

Import surveys into a list. Adds filename as a constant to each
element of the list.}
\usage{
read_surveys(import_file_names, .f = "read_rds", save_to_rds = TRUE)
}
\arguments{
\item{import_file_names}{A vector of file names to import.}

\item{.f}{A function to import the surveys with.
Defaults to \code{'read_rds'}. For SPSS files,
\code{read_spss} is recommended, which is a
well-parametrised version of \code{\link[haven]{read_spss}} that
saves some metadata, too.}

\item{save_to_rds}{Should it save the imported survey to .rds?
Defaults to \code{TRUE}.}
}
\value{
A list of the surveys.  Each element of the list is a data
frame. The respective file names are added to each data frame as a
constant column \code{filename}.
}
\description{
Read Survey Files

Import surveys into a list. Adds filename as a constant to each
element of the list.
}
\examples{
file1 <- system.file(
    "examples", "ZA7576.rds", package = "eurobarometer")
file2 <- system.file(
    "examples", "ZA5913.rds", package = "eurobarometer")

read_surveys (c(file1,file2), .f = 'read_rds' )
}
