% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retroharmonize-package.R
\docType{package}
\name{retroharmonize-package}
\alias{retroharmonize-package}
\title{retroharmonize: Ex Post Survey Data Harmonization}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Assist in reproducible retrospective (ex-post) harmonization of data, particularly individual level survey data, by providing tools for organizing metadata, standardizing the coding of variables, and variable names and value labels, including missing values, and documenting the data transformations, with the help of comprehensive s3 classes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://retroharmonize.dataobservatory.eu/}
  \item Report bugs at \url{https://github.com/dataobservatory-eu/retroharmonize/issues}
}

}
\author{
\strong{Maintainer}: Daniel Antal \email{daniel.antal@dataobservatory.eu} (\href{https://orcid.org/0000-0001-7513-6760}{ORCID})

Other contributors:
\itemize{
  \item Marta Kolczynska \email{mkolczynska@gmail.com} (\href{https://orcid.org/0000-0003-4981-0437}{ORCID}) [contributor]
}

}
\keyword{internal}
