% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_save_surveys.R
\name{subset_save_surveys}
\alias{subset_save_surveys}
\title{Subset and Save Surveys}
\usage{
subset_save_surveys(
  survey_list,
  selection_name = "trust",
  import_path = gesis_dir,
  export_path = "working"
)
}
\arguments{
\item{survey_list}{A vector of file names to import.}

\item{selection_name}{An identifier for the survey subset.}

\item{import_path}{The path to the survey files.}

\item{export_path}{The path where the subsets should be saved.}
}
\description{
Read a predefined survey list and variables.
}
