% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggest_var_names.R
\name{suggest_permanent_names}
\alias{suggest_permanent_names}
\title{Suggest Permanent Names}
\usage{
suggest_permanent_names(survey_program = "eurobarometer")
}
\arguments{
\item{survey_program}{Suggest permanent names for the survey progarm \code{"eurobarometer"}}
}
\value{
A character vector with suggested permanent names.
}
\description{
Suggest the use of established naming conventions.
}
\details{
Established survey programs usually have their own variable name conventions. 
The suggested constant names keep these variable names constant.
}
\examples{
suggest_permanent_names ( "eurobarometer" )
}
\seealso{
Other harmonization functions: 
\code{\link{collect_val_labels}()},
\code{\link{harmonize_na_values}()},
\code{\link{harmonize_values}()},
\code{\link{harmonize_var_names}()},
\code{\link{harmonize_waves}()},
\code{\link{label_normalize}()},
\code{\link{merge_waves}()},
\code{\link{na_range_to_values}()},
\code{\link{suggest_var_names}()}
}
\concept{harmonization functions}
