% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.R, R/survey_df.R
\name{is.survey_df}
\alias{is.survey_df}
\alias{survey_df}
\alias{print.survey_df}
\title{Create a survey data frame}
\usage{
is.survey_df(object)

survey_df(
  x,
  dataset_bibentry = NULL,
  dataset_subject = NULL,
  identifier,
  filename
)

is.survey_df(object)

\method{print}{survey_df}(x, ...)
}
\arguments{
\item{x}{A tibble or data frame that contains the survey data.}

\item{dataset_bibentry}{A list of bibliographic references and descriptive metadata
about the dataset as a whole created with 
\code{\link[dataset:datacite]{dataset::datacite}} or
\code{\link[dataset:dublincore]{dataset::dublincore}}.}

\item{dataset_subject}{The subject of the dataset; it can be a simple 
character string, or a more complex object created with 
\code{\link[dataset:subject]{dataset::subject}}.}

\item{filename}{The import file name.}

\item{id}{A mandatory identifier for the survey.}
}
\value{
A \code{\link{dataset::dataset_df}} object with with \code{id}, and   
\code{filename} metadata information.
}
\description{
Store the data of a survey in a tibble (data frame) with a unique
survey identifier, import filename, and optional document object identifier.
}
\details{
Whilst you can create a survey object with this helper function, 
it is most likely that you will receive it with an importing function, i.e. 
\code{\link{read_rds}}, \code{\link{read_spss}} \code{\link{read_dta}}, 
\code{\link{read_csv}} or 
their common wrapper \code{\link{read_survey}}.
}
\examples{
example_survey <- survey( 
  x = data.frame ( 
    rowid = 1:6,
    observations = runif(6)), 
  id = 'example', 
  filename = "no_file"
)
}
\seealso{
Other importing functions: 
\code{\link{survey}()}
}
\concept{importing functions}
